pyproject.toml
[project]
name = "mike-action-provider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "flask>=3.1.0",
    "globus-action-provider-tools>=0.20.0",
    "globus-cli>=3.35.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.41",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
]

[dependency-groups]
dev = [
    { include-group = "lint" },
]
lint = [
    "black>=25.1.0",
]
